created ui for a quick glance at the inventory
