--- frontend/src/styles/InventoryManagement.css
@@ -1,23 +1,0 @@
-.container {
-    display: "flex";
-    flex-direction: column;
-    align-content: center;
-    justify-content: center;
-    margin-bottom: 2rem;
-}
-
-.formContainer {
-    display: flex;
-    flex-direction: column;
-    align-content: center;
-    justify-content: center;
-}
-
-.formGroup {
-    margin: auto;
-    padding: 2rem;
-}
-
-.inventoryPaper {
-    padding: 2rem;
-}
