+++ frontend/src/styles/ManageInventory.css
@@ -1,0 +1,23 @@
+.container {
+    display: "flex";
+    flex-direction: column;
+    align-content: center;
+    justify-content: center;
+    margin-bottom: 2rem;
+}
+
+.formContainer {
+    display: flex;
+    flex-direction: column;
+    align-content: center;
+    justify-content: center;
+}
+
+.formGroup {
+    margin: auto;
+    padding: 2rem;
+}
+
+.inventoryPaper {
+    padding: 2rem;
+}
